requirements_utf8.txt
asgiref==3.9.1
channels>=4.3
channels-redis>=4.3
colorama==0.4.6
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-filter>=23.1
django-cors-headers>=4.0
drf-yasg>=1.21
celery>=5.3
redis>=4.5
psycopg2-binary>=2.9
msgpack>=1.1
inflection==0.5.1
packaging==25.0
pillow==11.3.0
PyJWT==2.10.1
pypng==0.20220715.0
pytz==2025.2
PyYAML==6.0.2
qrcode==7.4.2
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
requests
drf-spectacular灯湥祰汸㴾⸳⸱ਲ